sam_pdf/requirements.txt
# PDF Processing
PyMuPDF>=1.24.0
pymupdf4llm>=0.0.10
pdf2image>=1.17.0

# OCR
paddlepaddle>=2.6.0
paddleocr>=2.7.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# VLM Client
openai>=1.0.0
requests>=2.31.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
colorama>=0.4.6  # For colored diff output
pynput>=1.7.0    # For global hotkey detection (screenshot tool)

# Compression
pikepdf>=8.0.0


# OCR Benchmark Dependencies
# ==========================

# Core
PyMuPDF>=1.23.0
pdf2image>=1.16.0
Pillow>=10.0.0
numpy>=1.24.0

# PaddleOCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opt-einsum==3.3.0

# Hugging Face / Transformers
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.45.0
accelerate>=0.26.0
qwen-vl-utils>=0.0.4
tiktoken>=0.7.0
scipy>=1.11.0

# Nougat (Scientific OCR)
nougat-ocr>=0.1.17

# Optional for optimizations
bitsandbytes>=0.41.0  # For 8-bit/4-bit loading if GPU available
flash-attn>=2.0.0     # For faster inference if GPU available

# Local VLM HTTP server (OpenAI-compatible)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# (optional) smaller memory footprint
safetensors>=0.4.0
